requirements.txt
alembic==1.8.0
click==8.1.3
cycler==0.11.0
Flask==2.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.33.3
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
wtforms-html5==0.6.1
WTForms-SQLAlchemy==0.3
